* change polygon structure slightly and move "arc chopper" logic into the GDK   draw functions, and store only the arc info in the gerb_image struct. This   allows exporting with 100% accuracy, speeds up the cairo rendering, and reduces   the size of the gerb_image * update 4 golden files as a result of this change, since the cairo renderer now   renders with the default resolution instead of the mandated 1 degree resolution   with the previous method. * add more functionality to the export code, including info headers, polarity   changes, and finishing the polygon stuff. The only major code that still needs   to be added is macro exporting. This may wait until (or if) we re-arrange the   macro storage method. * change GDK renderer to use a double for the scale factor instead of an int. The int   was causing noticable error during rendering of large boards in comparison to   the cairo version (which was correct).

+++ src/draw-gdk.h
@@ -36,7 +36,7 @@
  */
 int 
 image2pixmap(GdkPixmap **pixmap, gerb_image_t *image, 
-	     int scale, double trans_x, double trans_y,
+	     double scale, double trans_x, double trans_y,
 	     enum polarity_t polarity);
 
 #endif /* DRAW_GDK_H */
@@ -53,7 +53,7 @@ image2pixmap(GdkPixmap **pixmap, gerb_image_t *image,
  */
 int gerbv_gdk_draw_amacro(GdkPixmap *pixmap, GdkGC *gc,
 		      instruction_t *program, unsigned int nuf_push,
-		      double *parameters, int scale, gint x, gint y);
+		      double *parameters, double scale, gint x, gint y);
 
 #endif /* DRAW_AMACRO_H */
